feat: move Sync Now to File menu with Cmd/Ctrl+S shortcut
Adds "Sync Now" as a proper File menu item (disabled when no workspace
is open) with the CmdOrCtrl+S accelerator. Removes the Sync Now button
and its associated result banner from the Workspace Peers dialog, per
issue #115.
